Переход на умные указатели

+++ lw1/Starovoytov_Alexey/MonteCarloPI/MathHelper.h
@@ -7,7 +7,7 @@ public:
 	template <unsigned int precision>
 	static double Round(double number)
 	{
-		unsigned int a = pow(10, precision);
+		double a = pow(10, precision);
 		return round(number * a) / a;
 	}
 };
